fix bug with permissions
